2020 Day 2

+++ 2020/01/main_test.go
@@ -8,11 +8,11 @@ import (
 
 // Examples of program input and output from https://adventofcode.com/2019/day/3
 func Test2(t *testing.T) {
-	a, b := findSum2([]int{1721, 979, 366, 299, 675, 1456}, 2020)
-	assert.Equal(t, a*b, 514579)
+	a, b := findSum2([]int64{1721, 979, 366, 299, 675, 1456}, 2020)
+	assert.Equal(t, a*b, int64(514579))
 }
 
 func Test3(t *testing.T) {
-	a, b, c := findSum3([]int{1721, 979, 366, 299, 675, 1456}, 2020)
-	assert.Equal(t, a*b*c, 241861950)
+	a, b, c := findSum3([]int64{1721, 979, 366, 299, 675, 1456}, 2020)
+	assert.Equal(t, a*b*c, int64(241861950))
 }
